feat: add eslint configurations
