+++ package.json
@@ -60,8 +60,8 @@
     "@changesets/changelog-github": "^0.4.8",
     "@changesets/cli": "^2.26.2",
     "@changesets/get-github-info": "^0.5.2",
-    "@fuels/eslint-config": "workspace:*",
-    "@fuels/changeset": "workspace:*"
+    "@fuels/changeset": "workspace:*",
+    "@fuels/eslint-plugin": "workspace:*"
   },
   "pnpm": {
     "peerDependencyRules": {
