+++ packages/eslint-plugin/package.json
@@ -1,23 +1,29 @@
 {
-  "name": "@fuels/eslint-config",
+  "name": "@fuels/eslint-plugin",
   "version": "0.0.2",
   "license": "Apache-2.0",
+  "access": "public",
   "main": "./index.js",
-  "publishConfig": {
-    "access": "public",
-    "files": [
-      "./index.js"
-    ]
+  "type": "commonjs",
+  "exports": {
+    ".": {
+      "default": "./index.js"
+    }
   },
+  "keywords": [
+    "eslint",
+    "eslintplugin",
+    "eslint-plugin",
+    "typescript"
+  ],
   "peerDependencies": {
     "eslint": "^8.44.0"
   },
   "dependencies": {
-    "@rushstack/eslint-patch": "^1.3.2",
+    "@next/eslint-plugin-next": "^13.4.9",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "eslint": "^8.44.0",
-    "eslint-config-next": "^13.4.9",
     "eslint-config-prettier": "^8.8.0",
     "eslint-import-resolver-node": "^0.3.7",
     "eslint-import-resolver-typescript": "^3.5.5",
